backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==1.10.15
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.142.0
beautifulsoup4==4.12.3
circle-user-controlled-wallets
web3==6.20.1
circleci
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1